refactor(cron): extract JobCard component; no behavior change
Lifts the per-job card JSX (header, actions, schedule summary, delete
confirmation, expandable run history) out of CronPanel.tsx into a
dedicated JobCard.tsx under cron/. JobCard accepts the prop shape the
rest of the Crons panel redesign will drive (onToggleExpand, onEdit,
onPauseResume, onRunNow, onDelete, onSelectRun) while preserving
existing visual layout, colors, and spacing.

Uses the shared ./formatters (formatSchedule, formatDuration,
formatAbsoluteTime) from Task 3 rather than re-implementing. Card root
now carries role="article" for the JobList test in Task 6. onSelectRun
is accepted but unused — Task 7 will wire run-row clicks.
