+++ apps/frontend/src/components/control/panels/CronPanel.tsx
@@ -6,25 +6,17 @@ import {
   Loader2,
   RefreshCw,
   Clock,
-  Play,
   Plus,
-  Pencil,
-  Trash2,
-  ChevronDown,
-  ChevronRight,
-  CheckCircle2,
-  XCircle,
-  MinusCircle,
   X,
 } from "lucide-react";
 import { useGatewayRpc, useGatewayRpcMutation } from "@/hooks/useGatewayRpc";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { cn } from "@/lib/utils";
+import { JobCard } from "./cron/JobCard";
 import type {
   CronJob,
   CronListResponse,
-  CronRunsResponse,
   CronSchedule,
   CronScheduleKind,
 } from "./cron/types";
@@ -58,44 +50,6 @@ const EMPTY_FORM: FormState = {
 };
 
 // --- Helpers ---
-
-function formatSchedule(schedule: CronSchedule): string {
-  if (!schedule || !schedule.kind) return "\u2014";
-  switch (schedule.kind) {
-    case "cron":
-      return schedule.expr ? `cron: ${schedule.expr}${schedule.tz ? ` (${schedule.tz})` : ""}` : "\u2014";
-    case "every": {
-      if (!schedule.everyMs) return "\u2014";
-      const ms = schedule.everyMs;
-      if (ms >= 86400000) return `Every ${Math.round(ms / 86400000)} day${ms >= 172800000 ? "s" : ""}`;
-      if (ms >= 3600000) return `Every ${Math.round(ms / 3600000)} hour${ms >= 7200000 ? "s" : ""}`;
-      return `Every ${Math.round(ms / 60000)} minute${ms >= 120000 ? "s" : ""}`;
-    }
-    case "at":
-      if (!schedule.at) return "\u2014";
-      try {
-        return `Once at ${new Date(schedule.at).toLocaleString()}`;
-      } catch {
-        return `Once at ${schedule.at}`;
-      }
-    default:
-      return JSON.stringify(schedule);
-  }
-}
-
-function formatDuration(ms: number): string {
-  if (ms < 1000) return `${ms}ms`;
-  if (ms < 60000) return `${(ms / 1000).toFixed(1)}s`;
-  return `${Math.round(ms / 60000)}m`;
-}
-
-function formatTimestamp(ms: number): string {
-  try {
-    return new Date(ms).toLocaleString();
-  } catch {
-    return String(ms);
-  }
-}
 
 function buildSchedule(form: FormState): CronSchedule {
   switch (form.scheduleKind) {
@@ -131,24 +85,6 @@ function jobToForm(job: CronJob): FormState {
     return { ...EMPTY_FORM, ...base, scheduleKind: "at", atDatetime };
   }
   return { ...EMPTY_FORM, ...base };
-}
-
-// --- Status badge ---
-
-function StatusBadge({ status }: { status?: "ok" | "error" | "skipped" }) {
-  if (!status) return null;
-  const config = {
-    ok: { icon: CheckCircle2, label: "OK", className: "text-[#2d8a4e]" },
-    error: { icon: XCircle, label: "Error", className: "text-red-500" },
-    skipped: { icon: MinusCircle, label: "Skipped", className: "text-yellow-500" },
-  };
-  const { icon: Icon, label, className } = config[status];
-  return (
-    <span className={cn("inline-flex items-center gap-1 text-xs", className)}>
-      <Icon className="h-3 w-3" />
-      {label}
-    </span>
-  );
 }
 
 // --- Create/Edit form ---
@@ -313,54 +249,6 @@ function CronJobForm({
   );
 }
 
-// --- Run history ---
-
-function RunHistory({ jobId }: { jobId: string }) {
-  const { data, error, isLoading } = useGatewayRpc<CronRunsResponse>("cron.runs", {
-    scope: "job",
-    id: jobId,
-    limit: 10,
-    sortDir: "desc",
-  });
-
-  if (isLoading) {
-    return (
-      <div className="flex items-center gap-2 py-2 text-xs text-[#8a8578]">
-        <Loader2 className="h-3 w-3 animate-spin" /> Loading history...
-      </div>
-    );
-  }
-
-  if (error) {
-    return <p className="text-xs text-destructive py-1">Failed to load history</p>;
-  }
-
-  const entries = data?.entries ?? [];
-  if (entries.length === 0) {
-    return <p className="text-xs text-[#8a8578] py-1">No runs yet</p>;
-  }
-
-  return (
-    <div className="space-y-1">
-      {entries.map((entry, i) => (
-        <div key={i} className="flex items-center gap-3 text-xs py-1 border-t border-[#e0dbd0]">
-          <StatusBadge status={entry.status} />
-          <span className="text-[#8a8578]">{formatTimestamp(entry.triggeredAtMs)}</span>
-          {entry.durationMs != null && (
-            <span className="text-[#8a8578]">{formatDuration(entry.durationMs)}</span>
-          )}
-          {entry.summary && (
-            <span className="text-[#5a5549] truncate flex-1">{entry.summary}</span>
-          )}
-          {entry.error && (
-            <span className="text-destructive truncate flex-1">{entry.error}</span>
-          )}
-        </div>
-      ))}
-    </div>
-  );
-}
-
 // --- Main panel ---
 
 export function CronPanel() {
@@ -372,7 +260,6 @@ export function CronPanel() {
   const [mode, setMode] = useState<"list" | "create" | "edit">("list");
   const [editingJob, setEditingJob] = useState<CronJob | null>(null);
   const [expandedJob, setExpandedJob] = useState<string | null>(null);
-  const [deletingJob, setDeletingJob] = useState<string | null>(null);
   const [saving, setSaving] = useState(false);
   const [feedback, setFeedback] = useState<{ type: "success" | "error"; message: string } | null>(null);
 
@@ -429,7 +316,6 @@ export function CronPanel() {
   const handleDelete = async (id: string) => {
     try {
       await callRpc("cron.remove", { id });
-      setDeletingJob(null);
       mutate();
       showFeedback("success", "Job deleted");
     } catch (err) {
@@ -547,135 +433,24 @@ export function CronPanel() {
 
       {mode === "list" && jobs.length > 0 && (
         <div className="space-y-2">
-          {jobs.map((job) => {
-            const isExpanded = expandedJob === job.id;
-            const isDeleting = deletingJob === job.id;
-
-            return (
-              <div key={job.id} className="rounded-lg border border-[#e0dbd0] overflow-hidden">
-                {/* Job header */}
-                <div className="p-3 space-y-2">
-                  <div className="flex items-center justify-between">
-                    <button
-                      className="flex items-center gap-2 text-left flex-1 min-w-0"
-                      onClick={() => setExpandedJob(isExpanded ? null : job.id)}
-                    >
-                      {isExpanded ? (
-                        <ChevronDown className="h-3.5 w-3.5 opacity-50 shrink-0" />
-                      ) : (
-                        <ChevronRight className="h-3.5 w-3.5 opacity-50 shrink-0" />
-                      )}
-                      <Clock className="h-3.5 w-3.5 opacity-50 shrink-0" />
-                      <span className="text-sm font-medium truncate">{job.name || job.id}</span>
-                      <span
-                        className={cn(
-                          "text-[10px] px-1.5 py-0.5 rounded-full shrink-0",
-                          job.enabled
-                            ? "bg-[#e8f5e9] text-[#2d8a4e]"
-                            : "bg-[#f3efe6] text-[#8a8578]",
-                        )}
-                      >
-                        {job.enabled ? "active" : "paused"}
-                      </span>
-                    </button>
-                    <div className="flex gap-1 shrink-0">
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => handleRun(job.id)}
-                        title="Run now"
-                      >
-                        <Play className="h-3.5 w-3.5" />
-                      </Button>
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => {
-                          setEditingJob(job);
-                          setMode("edit");
-                        }}
-                        title="Edit"
-                      >
-                        <Pencil className="h-3.5 w-3.5" />
-                      </Button>
-                      <Button
-                        variant={job.enabled ? "outline" : "default"}
-                        size="sm"
-                        onClick={() => handleToggle(job.id, !!job.enabled)}
-                      >
-                        {job.enabled ? "Disable" : "Enable"}
-                      </Button>
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => setDeletingJob(isDeleting ? null : job.id)}
-                        title="Delete"
-                        className="text-destructive/70 hover:text-destructive"
-                      >
-                        <Trash2 className="h-3.5 w-3.5" />
-                      </Button>
-                    </div>
-                  </div>
-
-                  {/* Schedule + last run info */}
-                  <div className="flex items-center gap-3 text-xs text-[#8a8578] pl-7">
-                    <span>{formatSchedule(job.schedule)}</span>
-                    {job.state?.lastRunStatus && (
-                      <>
-                        <span>&middot;</span>
-                        <StatusBadge status={job.state.lastRunStatus} />
-                      </>
-                    )}
-                    {job.state?.nextRunAtMs && (
-                      <>
-                        <span>&middot;</span>
-                        <span>Next: {formatTimestamp(job.state.nextRunAtMs)}</span>
-                      </>
-                    )}
-                  </div>
-
-                  {job.description && (
-                    <div className="text-xs text-[#5a5549] pl-7">{job.description}</div>
-                  )}
-                </div>
-
-                {/* Delete confirmation */}
-                {isDeleting && (
-                  <div className="px-3 pb-3">
-                    <div className="flex items-center justify-between rounded-md bg-destructive/10 border border-destructive/20 px-3 py-2">
-                      <span className="text-sm text-destructive">
-                        Delete &ldquo;{job.name || job.id}&rdquo;? This cannot be undone.
-                      </span>
-                      <div className="flex gap-1">
-                        <Button
-                          variant="outline"
-                          size="sm"
-                          onClick={() => setDeletingJob(null)}
-                        >
-                          Cancel
-                        </Button>
-                        <Button
-                          variant="destructive"
-                          size="sm"
-                          onClick={() => handleDelete(job.id)}
-                        >
-                          Delete
-                        </Button>
-                      </div>
-                    </div>
-                  </div>
-                )}
-
-                {/* Expanded: run history */}
-                {isExpanded && (
-                  <div className="px-3 pb-3 pl-7 border-t border-[#e0dbd0]">
-                    <p className="text-xs font-medium text-[#8a8578] pt-2 pb-1">Run History</p>
-                    <RunHistory jobId={job.id} />
-                  </div>
-                )}
-              </div>
-            );
-          })}
+          {jobs.map((job) => (
+            <JobCard
+              key={job.id}
+              job={job}
+              expanded={expandedJob === job.id}
+              onToggleExpand={() => setExpandedJob(expandedJob === job.id ? null : job.id)}
+              onEdit={() => {
+                setEditingJob(job);
+                setMode("edit");
+              }}
+              onPauseResume={() => handleToggle(job.id, !!job.enabled)}
+              onRunNow={() => handleRun(job.id)}
+              onDelete={() => handleDelete(job.id)}
+              onSelectRun={() => {
+                /* no-op for now; Task 7 wires */
+              }}
+            />
+          ))}
         </div>
       )}
     </div>
